refactor: extract shared functionality
