+++ src/node-lib/curriculum-api-2023/helpers/index.ts
@@ -1,3 +1,6 @@
+import { RawSyntheticUVLesson } from "../queries/lessonDownloads/rawSyntheticUVLesson.schema";
+import { lessonPathwaySchema } from "../shared.schema";
+
 export const toSentenceCase = (str: string) => {
   return str.charAt(0).toUpperCase() + str.slice(1).toLowerCase();
 };
@@ -17,4 +20,26 @@ export const isTierValid = (
   } else {
     return true;
   }
+};
+
+/**
+ * Used in canonical lesson pages to construct the pathways for a lesson
+ */
+export const constructPathwayLesson = (lesson: RawSyntheticUVLesson) => {
+  return lessonPathwaySchema.parse({
+    programmeSlug: lesson.programme_slug,
+    unitSlug: lesson.unit_data.slug,
+    unitTitle: lesson.programme_fields.optionality ?? lesson.unit_data.title,
+    keyStageSlug: lesson.programme_fields.keystage_slug,
+    keyStageTitle: toSentenceCase(lesson.programme_fields.keystage_description),
+    subjectSlug: lesson.programme_fields.subject_slug,
+    subjectTitle: lesson.programme_fields.subject,
+    lessonCohort: lesson.lesson_data._cohort,
+    examBoardSlug: lesson.programme_fields.examboard_slug,
+    examBoardTitle: lesson.programme_fields.examboard,
+    lessonSlug: lesson.lesson_slug,
+    lessonTitle: lesson.lesson_data.title,
+    tierSlug: lesson.programme_fields.tier_slug,
+    tierTitle: lesson.programme_fields.tier_description,
+  });
 };
