+++ src/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.query.ts
@@ -2,12 +2,8 @@ import {
   LessonOverviewQuery,
   Published_Mv_Synthetic_Unitvariant_Lessons_By_Keystage_10_0_0_Bool_Exp,
 } from "../../generated/sdk";
-import {
-  lessonOverviewQuizData,
-  LessonPathway,
-  lessonPathwaySchema,
-} from "../../shared.schema";
-import { toSentenceCase } from "../../helpers";
+import { lessonOverviewQuizData, LessonPathway } from "../../shared.schema";
+import { constructPathwayLesson, toSentenceCase } from "../../helpers";
 
 import lessonOverviewSchema, {
   lessonContentSchema,
@@ -124,23 +120,7 @@ export function getCopyrightContent(
 const getPathways = (res: LessonOverviewQuery): LessonPathway[] => {
   const pathways = res.browseData.map((l) => {
     const lesson = lessonBrowseDataByKsSchema.parse(l);
-    const unitTitle =
-      lesson.programme_fields.optionality ?? lesson.unit_data.title;
-    const pathway = {
-      programmeSlug: lesson.programme_slug,
-      unitSlug: lesson.unit_slug,
-      unitTitle,
-      keyStageSlug: lesson.programme_fields.keystage_slug,
-      keyStageTitle: lesson.programme_fields.keystage_description,
-      subjectSlug: lesson.programme_fields.subject_slug,
-      subjectTitle: lesson.programme_fields.subject,
-      lessonCohort: lesson.lesson_data._cohort,
-      examBoardSlug: lesson.programme_fields.examboard_slug,
-      examBoardTitle: lesson.programme_fields.examboard,
-      tierSlug: lesson.programme_fields.tier_slug,
-      tierTitle: lesson.programme_fields.tier_description,
-    };
-    return lessonPathwaySchema.parse(pathway);
+    return constructPathwayLesson(lesson);
   });
   return pathways;
 };
